Refactor exercise l ibrary.local.01.25

+++ lib/controllers/routine_user_controller.dart
@@ -54,7 +54,7 @@ class RoutineUserController extends ChangeNotifier {
   }
 
   double weight() {
-    return _amplifyRoutineUserRepository.user?.weight ?? 0;
+    return (_amplifyRoutineUserRepository.user?.weight ?? 0).toDouble();
   }
 
   void clear() {
